# src/express-server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: API REST para sistema de delivery - Trabalho DW1 UMAIA 2025/26
  contact:
    name: inf25dw1g13
    email: a046756@umaia.pt
components:
  schemas:
...
    Prato:
      title: Prato
      type: object
      required:
        - restaurante_id
        - nome
        - preco
      additionalProperties: false
      properties:
        id:
          type: integer
          readOnly: true
          description: ID do prato (gerado automaticamente)
        restaurante_id:
          type: integer
          description: ID do restaurante a que pertence o prato
          example: 1
        nome:
          type: string
          description: Nome do prato
          example: "Pizza Margherita"
        descricao:
          type: string
          description: "(Opcional) Descrição do prato"
          example: "Pizza clássica com tomate, mozzarella e manjericão"
        preco:
          type: number
          format: float
          description: "Preço do prato em euros"
          example: 8.50
        vegetariano:
          type: boolean
          default: false
          description: "(Opcional) Se o prato é vegetariano (padrão: false)"
          example: true
        disponivel:
          type: boolean
          default: true
          description: "(Opcional) Se o prato está disponível (padrão: true)"
          example: true
...
    Ingrediente:
      title: Ingrediente
      type: object
      required:
        - nome
      additionalProperties: false
      properties:
        id:
          type: integer
          readOnly: true
          description: ID do ingrediente (gerado automaticamente)
        nome:
          type: string
          minLength: 1
          maxLength: 100
          description: Nome do ingrediente (deve ser único)
          example: "Tomate"
        unidade:
          type: string
          enum: [g, kg, ml, l, unidade, fatia]
          default: g
          description: "(Opcional) Unidade de medida (padrão: g)"
          example: "g"
        alergeno:
          type: boolean
          default: false
          description: "(Opcional) Se o ingrediente é um alergénio (padrão: false)"
          example: false
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Data de criação do ingrediente
...
    Error:
      title: Error
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      # Schema é reutilizado em várias respostas de erro .
      # As respostas usam $ref para apontar para este modelo, todas as mensagens ficam com o mesmo formato
  examples:
...
    PratoExample01:
      value:
        restaurante_id: 1
        nome: "Pizza Margherita"
        descricao: "Pizza clássica com tomate, mozzarella e manjericão"
        preco: 8.50
        vegetariano: true
        disponivel: true
    PratoExample02:
      value:
        restaurante_id: 2
        nome: "Burger Clássico"
        descricao: "Hambúrguer de carne, queijo, alface, tomate e molho especial"
        preco: 6.90
        vegetariano: false
        disponivel: true
...
    IngredienteExample01:
      value:
        nome: "Tomate"
        unidade: "g"
        alergeno: false
    IngredienteExample02:
      value:
        nome: "Amendoim"
        unidade: "g"
        alergeno: true
...
paths:
...
  /api/pratos:
    get:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: getPratos
      summary: Listar pratos
      parameters:
        - name: restaurante_id
          in: query
          schema:
            type: integer
        - name: categoria_id
          in: query
          schema:
            type: integer
        - name: vegetariano
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de pratos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prato'
    post:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: createPrato
      summary: Criar prato
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prato'
            examples:
              PratoExample01:
                $ref: '#/components/examples/PratoExample01'
              PratoExample02:
                $ref: '#/components/examples/PratoExample02'
      responses:
        '201':
          description: Prato criado
  /api/pratos/{id}:
    get:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: getPrato
      summary: Obter prato por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Retrieve Prato id 1
            two:
              value: 2
              summary: Retrieve Prato id 2
      responses:
        '200':
          description: Prato encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prato'
    put:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: updatePrato
      summary: Atualizar prato
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Update Prato id 1
            two:
              value: 2
              summary: Update Prato id 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prato'
            examples:
              PratoExample01:
                $ref: '#/components/examples/PratoExample01'
              PratoExample02:
                $ref: '#/components/examples/PratoExample02'
      responses:
        '200':
          description: Atualizado
    delete:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: deletePrato
      summary: Apagar prato
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Delete Prato id 1
            two:
              value: 2
              summary: Delete Prato id 2
      responses:
        '204':
          description: Removido
  /api/pratos/{id}/ingredientes:
    get:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: getIngredientesPrato
      summary: Listar ingredientes de um prato (Relação m:n)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Lista de ingredientes do prato
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nome:
                      type: string
                    unidade:
                      type: string
                    quantidade:
                      type: number
                    obrigatorio:
                      type: boolean
    post:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: addIngredientePrato
      summary: Adicionar ingrediente a um prato (Relação m:n)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ingrediente_id
                - quantidade
              properties:
                ingrediente_id:
                  type: integer
                  description: ID do ingrediente
                quantidade:
                  type: number
                  description: Quantidade do ingrediente
                obrigatorio:
                  type: boolean
                  default: true
                  description: Se o ingrediente é obrigatório
      responses:
        '201':
          description: Ingrediente adicionado ao prato
    delete:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: removeIngredientePrato
      summary: Remover ingrediente de um prato (Relação m:n)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID do prato
        - name: ingrediente_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID do ingrediente a remover
      responses:
        '204':
          description: Ingrediente removido do prato com sucesso (sem conteúdo no body)
        '404':
          description: Prato ou ingrediente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/ingredientes:
    get:
      tags: [Ingredientes]
      x-eov-operation-handler: controllers/IngredientesController
      operationId: getIngredientes
      summary: Listar ingredientes
      parameters:
        - name: alergeno
          in: query
          schema:
            type: boolean
          description: Filtrar por ingredientes alergénios (opcional)
          example: true
      responses:
        '200':
          description: Lista de ingredientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingrediente'
    post:
      tags: [Ingredientes]
      x-eov-operation-handler: controllers/IngredientesController
      operationId: createIngrediente
      summary: Criar ingrediente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ingrediente'
            examples:
              IngredienteExample01:
                $ref: '#/components/examples/IngredienteExample01'
              IngredienteExample02:
                $ref: '#/components/examples/IngredienteExample02'
      responses:
        '201':
          description: Ingrediente criado
...